Refactor `lib.rs` into `choice`, `ctoption`, and `traits` submodules
There is a lot of functionality already in `lib.rs`.

Refactoring into submodules helps you find the specific types (or
traits) that you are looking for.

This also refactors the tests so the relevant tests for any of the three
things above can be found in a conditionally-included `tests` submodule
so you don't have to change files to find the tests, and only see tests
relevant to what is being defined in a particular submodule.
